Moved gpgsync.py logic directly into __init__.py, and added cli flag for debug mode
